# aspnetcore-server-generated/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Calculator API",
    "description" : "REST API that performs a math operation on two numbers via POST. The operation is provided in the `X-Operation` header (add, subtract, multiply, divide). Secured with Bearer JWT. Includes an Auth endpoint to mint a short-lived token for testing.\n",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5029",
    "description" : "Local HTTP (launchSettings http profile)"
  }, {
    "url" : "https://localhost:7008",
    "description" : "Local HTTPS (launchSettings https profile)"
  } ],
  "tags" : [ {
    "name" : "Auth",
    "description" : "Helper endpoints for authentication (testing only)."
  }, {
    "name" : "Calculator",
    "description" : "Calculator operations."
  } ],
  "paths" : {
    "/api/auth/token" : {
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "Issue a JWT token for testing",
        "description" : "Returns a short-lived JWT (includes `exp`) signed with your app key.",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "Optional username to include as the `sub` claim.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "default" : "demo"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Token issued",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                },
                "examples" : {
                  "ok" : {
                    "value" : {
                      "token" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                      "expiresAt" : "2025-09-16T21:45:00.000Z"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/calculator" : {
      "post" : {
        "tags" : [ "Calculator" ],
        "summary" : "Perform a math operation",
        "description" : "Executes the operation given by the `X-Operation` header on two numbers provided in the request body.\n",
        "parameters" : [ {
          "name" : "X-Operation",
          "in" : "header",
          "description" : "Operation to perform",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "add", "subtract", "multiply", "divide" ]
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OperationRequest"
              },
              "examples" : {
                "add" : {
                  "summary" : "Add two numbers",
                  "value" : {
                    "number1" : 10,
                    "number2" : 5
                  }
                },
                "divide" : {
                  "summary" : "Divide two numbers",
                  "value" : {
                    "number1" : 20,
                    "number2" : 4
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful calculation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_1"
                },
                "examples" : {
                  "ok" : {
                    "value" : {
                      "result" : 15
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request (missing header / invalid operation / division by zero)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "examples" : {
                  "missingHeader" : {
                    "value" : {
                      "error" : "Missing required header: X-Operation"
                    }
                  },
                  "invalidOp" : {
                    "value" : {
                      "error" : "Invalid operation: pow"
                    }
                  },
                  "divZero" : {
                    "value" : {
                      "error" : "Division by zero"
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized (missing/invalid bearer token)"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "OperationRequest" : {
        "required" : [ "number1", "number2" ],
        "type" : "object",
        "properties" : {
          "number1" : {
            "type" : "number",
            "description" : "First number"
          },
          "number2" : {
            "type" : "number",
            "description" : "Second number"
          }
        }
      },
      "Error" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "type" : "string"
          }
        }
      },
      "inline_response_200" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string",
            "description" : "The bearer token"
          },
          "expiresAt" : {
            "type" : "string",
            "description" : "UTC expiration time",
            "format" : "date-time"
          }
        },
        "example" : {
          "expiresAt" : "2000-01-23T04:56:07.000+00:00",
          "token" : "token"
        }
      },
      "inline_response_200_1" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "number",
            "description" : "The numeric result of the operation"
          }
        },
        "example" : {
          "result" : 0.8008281904610115
        }
      }
    },
    "parameters" : {
      "XOperation" : {
        "name" : "X-Operation",
        "in" : "header",
        "description" : "Operation to perform",
        "required" : true,
        "style" : "simple",
        "explode" : false,
        "schema" : {
          "type" : "string",
          "enum" : [ "add", "subtract", "multiply", "divide" ]
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}